Add file error handling and warning for portset.
Program will now gracefully print an error when unsuccessfully trying to open capture, dump, and replay files.

The `portset` command will now print a large WARNING if you set both port A and B to the same device (e.g. /dev/ttyUSB0).  It won't throw a full error, as I wasn't sure if there could be a legit case for alloeing this (and if we throw an error, it could get confusing/frustrating to the user if they are trying to swap port devices between the two ports, where allowing the same device name --but warning about it-- lets them swap devices names without quitting the program or locating a third device name to use temporarily).